pyproject.toml
[project]
name = "backend-oauth2-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Nuttapat-Office",email = "nuttapat.c@swiftdynamics.co.th"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "django (==5.2.1)",
    "djangorestframework (==3.16.0)",
    "python-keycloak (==5.5.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
